共通処理を基底クラスに移動
