+++ themes/turn-battle/imai/turn-battle.py
@@ -1,5 +1,7 @@
-import Brave, Enemy
+from Brave import Brave
+from Enemy import Enemy
 from enum import Enum
+import random
 
 class BattleState(Enum):
     BEFORE＿BATTLE = 0
@@ -14,6 +16,7 @@ class TurnBattle():
 
     def turn_start(self, brave:Brave, enemy:Enemy)->None:
         self.battle_state = BattleState.BATTLE_START
+        brave.self_introduction()
         enemy.self_introduction()
         enemy.self_status()
         brave.self_status()
@@ -26,8 +29,6 @@ class TurnBattle():
             print(f"TURN {self.turn_num}")
             self.brave_turn(brave, enemy)
             self.enemy_turn(brave, enemy)
-            enemy.self_status()
-            brave.self_status()
             self.battle_state = self.battle_result_check(brave, enemy)
 
     
@@ -47,18 +48,21 @@ class TurnBattle():
             print("相打ち…")
             print("----------------------------------------------")
             print("そういうこともあるよ")
+            print("----------------------------------------------")
             return BattleState.BATTLE_RESULT
         elif not is_brave_arive:
             print("----------------------------------------------")
             print(f"{enemy.get_name()}の勝利")
             print("----------------------------------------------")
             print("おお 勇者！ しんでしまうとは なにごとだ！")
+            print("----------------------------------------------")
             return BattleState.BATTLE_RESULT
         elif not is_enemy_arive:
             print("----------------------------------------------")
             print(f"勇者:{brave.get_name()}の勝利")
             print("----------------------------------------------")
             print("今日も世界に平和が訪れた")
+            print("----------------------------------------------")
             return BattleState.BATTLE_RESULT
 
         return BattleState.IN_BATTLE
@@ -67,12 +71,10 @@ turn_battle = TurnBattle()
  # 敵味方の生成・出力
 print("勇者の名前を入力してね。")
 name = input()
-brave = Brave.Brave(name)
-brave.self_introduction()
-enemy = Enemy.Enemy()
+brave = Brave(str(name))
+enemy = Enemy(random.choice(["スライム", "ゴブリン", "ドラゴン"]))
 
 # 開始
 turn_battle.turn_start(brave, enemy)
 turn_battle.battle_turn(brave, enemy)
-print("----------------------------------------------")
